Extract, convert parseLesson to TS
